hahsltd-main/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0 